Fix handling of ping timeouts in broker
This commit fixes a bug where the server would ignore the client's
declared keep alive interval.  Clients which failed to send a PING
within the timeout were tolerated although proper mechanisms to drop
their connections seemed to be implemented.

+++ src/PicoMQTT/connection.h
@@ -63,7 +63,7 @@ class Connection {
         virtual void on_disconnect();
 
         ClientWrapper client;
-        uint16_t keep_alive_millis;
+        unsigned long keep_alive_millis;
 
         virtual void handle_packet(IncomingPacket & packet);
